Move show() method into Plotting.py.
