+++ QGL/__init__.py
@@ -1,9 +1,9 @@
 from Channels import QubitFactory, MeasFactory, Qubit
 from PulsePrimitives import *
 from Compiler import compile_to_hardware
-from PulseSequencer import show, align, build_waveforms, plot_waveforms
+from PulseSequencer import align, build_waveforms, plot_waveforms
 from ControlFlow import repeat, repeatall, qif, qwhile, qdowhile, qfunction, qwait, qsync
 from BasicSequences import *
-from Plotting import output_file, output_notebook
+from Plotting import output_file, output_notebook, show
 from PulseSequencePlotter import plot_pulse_files
 from Tomography import state_tomo, process_tomo
